pyproject.toml
[project]
name = "inferencing-maven-assignments"
version = "0.1.0"
description = ""
authors = [
    {name = "Piotr Goral",email = "piotrek.grl@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.129.2",
    "uvicorn>=0.41.0",
    "httpx>=0.28.1",
    "pyyaml>=6.0.2",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[dependency-groups]
dev = [
    "black (>=26.1.0,<27.0.0)"
]

[tool.black]
line-length = 88
target-version = ["py313"]